- Clean-up which FID is used for API calls to Raven Colonial - Update FormPlayDev to work against a singular `devQuest` - Support `on_Status` - Support routes rendered on human settlement maps - Reduce excess decimal places in PlotStationInfo

+++ SrvSurvey/net/spansh.cs
@@ -1182,7 +1182,7 @@ namespace SrvSurvey.net
                 public string controllingFaction;
                 public string controllingFactionState;
                 public double distanceToArrival;
-                public Dictionary<string, float>? economies;
+                public Dictionary<string, double>? economies;
                 public string government;
                 public long id;
                 public LandingPads landingPads;
